+++ dependent-literals-plugin/CHANGELOG.md
@@ -1,3 +1,7 @@
 # 0.1.0.0
 
 Initial version.
+
+# 0.1.0.1
+
+* Updated tests to use `Wrapped Num` rather than `StockLit`.