+++ dependent-literals/CHANGELOG.md
@@ -1,3 +1,7 @@
 # 0.1.0.0
 
 Initial version.
+
+# 0.1.1.0
+
+* Moved the `StockLit` instance to `Wrapped Num` (kept `StockLit` as a newtype).